+++ index.js
@@ -1,9 +1,9 @@
 import express from "express";
 import cors from "cors";
-import connection from "./config/database.js";
-import accountsRouter from "./routes/accounts.js";
-import categoriesRouter from "./routes/categories.js";
-import authRouter from "./routes/auth.js";
+import connection from "./config/database.config.js";
+import accountsRouter from "./routes/accounts.route.js";
+import categoriesRouter from "./routes/categories.route.js";
+import authRouter from "./routes/auth.route.js";
 
 const app = express();
