+++ routes/auth.route.js
@@ -1,5 +1,5 @@
 import express from "express";
-import { login } from "../services/auth.js";
+import { login } from "../services/auth.service.js";
 const authRouter = express.Router();
 
 authRouter.post("/login", (req, res) => {
